feat(http): move to rpc package
