+++ transport/http/module.go
@@ -3,14 +3,14 @@ package http
 import (
 	"net/http"
 
-	nh "github.com/alexfalkowski/go-service/net/http"
+	"github.com/alexfalkowski/go-service/net/http/rpc"
 	"go.uber.org/fx"
 )
 
 // Module for fx.
 var Module = fx.Options(
 	fx.Provide(http.NewServeMux),
-	fx.Invoke(nh.Register),
+	fx.Invoke(rpc.Register),
 	fx.Provide(NewServer),
 	fx.Invoke(RegisterMetrics),
 )
